refactor : search, pagination
