+++ src/components/post/pagination/index.scss
@@ -1,6 +1,6 @@
 @import '@scss/abstracts/mixin';
 
-.post-list-page-box {
+.posts-pagination {
   display: flex;
   justify-content: space-between;
   align-items: center;
