+++ src/components/search/index.scss
