+++ src/components/search/recommend/index.scss
@@ -1,4 +1,4 @@
-.recommend-box {
+.recommend-post-tag {
   margin-top: 12px;
   display: flex;
   align-items: center;
